modular structure implimented
