+++ src/cart/add-item.js
@@ -1,0 +1,15 @@
+import client from "../shopify-client";
+
+function addItem(id) {
+  const checkoutId = localStorage.getItem("sandy-shop-checkoutId");
+  const lineItems = {
+    variantId: id,
+    quantity: 1
+  };
+
+  return client.checkout
+    .addLineItems(checkoutId, lineItems)
+    .catch(err => console.log("Problem adding item to cart", err));
+}
+
+export default addItem;
