+++ src/cart/close.js
@@ -1,0 +1,6 @@
+function close() {
+  // close the side menu
+  document.querySelector("#side-menu").style.width = "0";
+}
+
+export default close;
