+++ src/cart/index.js
@@ -1,0 +1,8 @@
+import renderCart from "./render-cart";
+import open from "./open";
+import close from "./close";
+import addItem from "./add-item";
+import addItemAndCheckout from "./add-item-and-checkout";
+import remove from "./remove";
+
+export default { renderCart, open , close, addItem , addItemAndCheckout , remove };
