+++ src/cart/open.js
@@ -1,0 +1,6 @@
+function open() {
+  // reveal the side menu.
+  document.querySelector("#side-menu").style.width = "400px";
+}
+
+export default open;
