+++ src/product-actions/fetch-feat-collect.js
@@ -1,10 +1,8 @@
-import client from "./shopify-client";
+import client from "../shopify-client";
 
 function fetchFeaturedCollection(id) {
   // Fetch a single collection by ID, including its products
-  return client.collection
-    .fetchWithProducts(id)
-    .catch(err => console.log(err));
+  return client.collection.fetchWithProducts(id).catch(err => console.log(err));
 }
 
 export default fetchFeaturedCollection;
