+++ src/product-actions/fetch-feat-item.js
@@ -1,13 +1,9 @@
-import client from "./shopify-client";
+import client from "../shopify-client";
 
 function fetchFeaturedItem(id) {
     // Fetch a single product by ID
     return client.product
       .fetch(id)
-    //   .then(product => {
-    //     // Do something with the product
-    //     displayFeaturedItem(product);
-    //   })
       .catch(err => console.log(err));
   }
