+++ src/product-actions/index.js
@@ -1,9 +1,8 @@
 import fetchAll from "./fetchAll";
-import removeProduct from "./removeP";
 import fetchFeaturedCollection from "./fetch-feat-collect";
 import fetchCollection from "./fetch-collection";
 import fetchFeaturedItem from "./fetch-feat-item";
 import fetchProduct from "./fetch-product";
 
 
-export default { fetchAll, removeProduct , fetchFeaturedCollection , fetchCollection , fetchFeaturedItem , fetchProduct };
+export default { fetchAll, fetchFeaturedCollection , fetchCollection , fetchFeaturedItem , fetchProduct  };
