+++ src/render/index.js
@@ -1,5 +1,7 @@
 import displayAll from "./display-all";
 import displayCollection from "./display-collection"; 
+import displayProduct from "./display-product";
+import displayFeaturedProduct from "./display-feat-product";
 
 
-export default { displayAll , displayCollection };
+export default { displayAll , displayCollection , displayProduct , displayFeaturedProduct };
